Move `getTestTranspile` into test helpers
