+++ src/files/BrsFile.Class.spec.ts
@@ -1,14 +1,13 @@
 import * as sinonImport from 'sinon';
 
 import { Program } from '../Program';
-import { getTestTranspile } from './BrsFile.spec';
 import type { BrsFile } from './BrsFile';
 import { expect } from 'chai';
 import { DiagnosticMessages } from '../DiagnosticMessages';
 import type { Diagnostic } from 'vscode-languageserver';
 import { Range, DiagnosticSeverity } from 'vscode-languageserver';
 import { ParseMode } from '../parser/Parser';
-import { expectZeroDiagnostics } from '../testHelpers.spec';
+import { expectZeroDiagnostics, getTestTranspile } from '../testHelpers.spec';
 import { standardizePath as s } from '../util';
 import * as fsExtra from 'fs-extra';
 import { TranspileState } from '../parser/TranspileState';
