+++ src/testHelpers.spec.ts
@@ -5,6 +5,11 @@ import { createSandbox } from 'sinon';
 import { expect } from 'chai';
 import type { CodeActionShorthand } from './CodeActionUtil';
 import { codeActionUtil } from './CodeActionUtil';
+// eslint-disable-next-line @typescript-eslint/consistent-type-imports
+import { BrsFile } from './files/BrsFile';
+import type { Program } from './Program';
+import { standardizePath as s } from './util';
+
 /**
  * Trim leading whitespace for every line (to make test writing cleaner
  */
@@ -100,4 +105,54 @@ export function expectCodeActions(test: () => any, expected: CodeActionShorthand
         delete arg.diagnostics;
     }
     expect(args).to.eql(expected);
+}
+
+export function getTestTranspile(scopeGetter: () => [Program, string]) {
+    return (source: string, expected?: string, formatType: 'trim' | 'none' = 'trim', pkgPath = 'source/main.bs', failOnDiagnostic = true) => {
+        let [program, rootDir] = scopeGetter();
+        expected = expected ? expected : source;
+        let file = program.addOrReplaceFile<BrsFile>({ src: s`${rootDir}/${pkgPath}`, dest: pkgPath }, source);
+        program.validate();
+        let diagnostics = file.getDiagnostics();
+        if (diagnostics.length > 0 && failOnDiagnostic !== false) {
+            throw new Error(
+                diagnostics[0].range.start.line +
+                ':' +
+                diagnostics[0].range.start.character +
+                ' ' +
+                diagnostics[0]?.message
+            );
+        }
+        let transpiled = file.transpile();
+
+        let sources = [transpiled.code, expected];
+        for (let i = 0; i < sources.length; i++) {
+            if (formatType === 'trim') {
+                let lines = sources[i].split('\n');
+                //throw out leading newlines
+                while (lines[0].length === 0) {
+                    lines.splice(0, 1);
+                }
+                let trimStartIndex = null;
+                for (let lineIndex = 0; lineIndex < lines.length; lineIndex++) {
+                    //if we don't have a starting trim count, compute it
+                    if (!trimStartIndex) {
+                        trimStartIndex = lines[lineIndex].length - lines[lineIndex].trim().length;
+                    }
+                    //only trim the expected file (since that's what we passed in from the test)
+                    if (lines[lineIndex].length > 0 && i === 1) {
+                        lines[lineIndex] = lines[lineIndex].substring(trimStartIndex);
+                    }
+                }
+                //trim trailing newlines
+                while (lines[lines.length - 1]?.length === 0) {
+                    lines.splice(lines.length - 1);
+                }
+                sources[i] = lines.join('\n');
+
+            }
+        }
+        expect(trimMap(sources[0])).to.equal(sources[1]);
+        return transpiled;
+    };
 }
